pyproject.toml
[tool.poetry]
name = "dynamic-ip-provider-service"
version = "0.1.0"
description = ""
authors = ["MaximPyanin <maximpyanin@icloud.com>"]
readme = "README.md"
packages = [{include = "provider"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.112.0"
pymongo = "^4.8.0"
pydantic = {extras = ["email"], version = "^2.8.2"}
pre-commit = "^3.8.0"
ruff = "^0.5.6"
bcrypt = "^4.2.0"
cryptography = "^43.0.0"
pyjwt = {extras = ["extras"], version = "^2.9.0"}
apscheduler = "^3.10.4"
sendgrid = "^6.11.0"
uvicorn = "^0.30.5"
gunicorn = "^22.0.0"
dependency-injector = "^4.41.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
